requirements.txt
fastapi[all]
sqlalchemy[asyncio]
asyncpg
passlib[bcrypt]
python-jose